Support both legacy and new authentication endpoints (#9)
Starting with the Poppy release of FOLIO a new endpoint was introduced for authentication authn/login-with-expiry. The existing, legacy endpoint authn/login will remain in place until the Sunflower release. More information about this change can be found in the FOLIO Wiki: https://folio-org.atlassian.net/wiki/spaces/FOLIJET/pages/1396980/Refresh+Token+Rotation+RTR

LDLite was hardcoded to use the legacy endpoint. This change adds a flag for LDLite to use either the legacy or the new endpoint. Because LDLite already supported reauthenticating the only change needed was to update the endpoint. This change is intended to be 100% backwards compatible while allowing for users to test using LDLite with the new authentication pattern.

Additionally, because authentication tokens now expire the connect_okapi_token endpoint will stop functioning in the Sunflower release. It has been deprecated.