Use pdm for package management and build backend (#15)
The dependencies are hard-pinned and need upgrading. To make this process easier I chose to use pdm over pip to manage dependencies. A side effect of this change is using pdm for packaging instead of setuptools. PDM was chosen because I personally like it and find it easier to use than other python project/package management tools.

After setting up PDM I had troubles installing/locking the dependencies. The version of pandas that was pinned made installing numpy impossible because it conflicted with the version required for duckdb. Duckdb is actually used in LDLite so I unpinned the pandas dependency to let it resolve. Even though a version of numpy could be resolved it could not be built because the python version I was using was 3.13 and incompatible with such old versions of libraries. I ended up constraining the python to 3.9 to get everything installable. This version is going to be end of life in October 2025 and will have to be bumped with the rest of the dependencies.

Additionally, the dependency on psycopg2 required having postgres installed which I did not. I created an environment.yaml file for the conda environment manager. While using conda is not required for LDLite it is a useful collection of all "globally" required dependencies to do development. For example, the python and pdm versions are specified there as well.

Before bumping dependencies LDLite needs to have test coverage. To prepare for this I installed and configure pytest to be used to test both the behavior and the package importability.