Lint, format, and typecheck examples and tests (#18)
This PR is already getting huge and I wanted to isolate the changes to the actual package in another PR. There's no real behavior changes here and it is only in non-production code. (folio_demo.py has one minor change with a new str() call)

Ruff is the modern linter/formatter compared to black or pylint/flake etc. I'm pretty sure that the pylintrc file was added but not used. I removed it anyways.

Mypy is still more or less the easiest type checker to install and configure.