Lint and fix the main codebase (#19)
This PR has no changes to behavior (hopefully, all the test still pass). All rules for the Ruff linter have been enabled as well as using the Ruff formatter. There were a small number of Ruff lint rules that had to be turned off in the following categories:
* Complexity and performance issues that would require bigger refactors
* String concatenation because, for now, it's ddl manipulation

I was going to make mypy happy in this PR as well but it was already getting huge.