Strictly type the main library (#20)
Mypy was disabled for the main library. This enables it and also makes ruff warn on Any type annotations. I relied on Any and Unknown quite a bit to make ruff happy while linting and this PR introduces a number of types to replace the Any type. For the most part there should be no behavior changes with two exceptions:
* drop_json_tables method was broken for all dbs
* uuid columns were broken for postgres

Indices on postgres remain broken because it is a much bigger behavior change.