Add additional tests, fix uncovered bugs, and configure code coverage (#22)
This refactors the existing test case infrastructure to share more code. Additional test cases were added for the drop_tables and export_csv methods.

There were a few bugs uncovered and fixed with this PR:
* Dropping tables did not work for sqlite
* LDLite.select output format was off, I broke this earlier
* Datatypes were inconsistent between database types in the csv export
* Columns were inconsistent between database types in the csv export

Fixing the csv bugs is a (small but no less) breaking change.

coverage.py is enabled and configured. I chose not to cover anything excel related as I hope to remove that functionality. We're hitting around 80% with what I (somewhat self-congratulatory) think are high-quality tests. I'm not a huge fan of code coverage as a metric but it is helpful to know what is or isn't scary to change at a glance.