Switch to pylock and relax dependencies (#23)
pylock.toml is a newer pep to share lock files across package managers. This replaces the pdm.lock with the pylock.toml file. I've relaxed the hard pins on the dependencies, I'm not sure why they were hard pinned in the first place but the tests all pass with the upgraded dependency versions. I've also added a "minimal" version lock file to make sure to run tests against what we claim the lower dependency bound to be.