Replace requests with httpx and orjson (#25)
Roughly 1/3 of LDLite's runtime is spend downloading data from FOLIO. This PR replaces the existing code which used limit/offset paging with id based paging which should result in a significant speed up of the download portion. Switching the type of paging should also increase reliability/stability for larger tables like instances and holdings. I also changed all json operations to use orjson which is much more performant.

I changed from requests to httpx at the same time. It's a little more stable and performant out of the box and has async support to use in the future.

LDLite has been capable of downloading the source records but the reliability is spotty. This PR does some hardcoding to use the stream endpoint for source records without callers having to know or update their code.

Lastly, the existing code did not work with some endpoints because they use perPage instead of limit as the parameter name. This PR fixes #5. I'm not currently aware of other "non-compliant" endpoints.

Some light refactoring allowed pulling out the dowloading code into a separate module which can be tested independently and allowed for confidence in the srs and perPage changes.