Deprecate sqlite and slsx functionality (#27)
For the transformation performance refactor I'm planning to use the native json functionality of duckdb and postgres. Sqlite has it but at a later version than what is built into python. It has to be removed to avoid maintaining two transformation code paths.

xlsx functionality is untested and it would be nice to drop it and the dependency. Excel can open csvs.