Re-enable indexing for postgres databases (#31)
When experimental sqlite support was added the indexing on postgres was accidentally removed. When I discovered this bug it didn't seem like anyone was in a hurry to have it fixed so I wanted to rethink it before adding the behavior back in. For this PR, we're only indexing "id" and "id like" columns which are

* Columns literally __id and id
* Columns that act as indices for lists ending in __o
* Columns that end in "_id" which are probably foreign keys
* UUID columns which are probably foreign keys

While this isn't perfect (or final) I think it strikes a good balance between re-enabling the existing behavior that indexed all columns (which took up a bunch of disk space unnecessarily) and have the indexes completely gone (which has worse performance).

This will close [#16](https://github.com/library-data-platform/ldlite/issues/16) when released.