Re-implement dictionary queries and add more robust default query handling (#33)
While working on the fix to the source record endpoint I realized I dropped the ability to pass dictionaries as queries at some point. The original behavior was really only documented in the srs.md file. I've re-added the functionality as well as types and tests.

Previously, `cql.allRecords sortBy id asc` or some variation of this query was everywhere in the docs. This query can still be passed but it won't have an effect as we've switched to id based paging from offset based paging. I had done this less robustly before but wanted to flesh it out more with tests and everything.