Switch to httpx-folio (#37)
There was a lot of folio/httpx logic piling up here that wasn't specifically related to ldlite. This was likely to increase over time with ECS and async support. I moved the authentication and query parameter code to a separate httpx-folio library which I then switched to with this PR.

Coincidentally, there were some edge cases around sorting and paging that weren't well handled by LDLite and the intended behavior was never clear either. This became worse when I switched to id offset based paging with 3.1. Pulling the query parameter logic out allowed for more thorough testing and handling of these edge cases and LDLite should behave better for ad-hoc usage where queries are more likely to be sorted and subsets of data are more likely to be returned.

I also accidentally exposed the new folio module in 3.1 when I intended for it to be internal.