Abstract ingestion database interactions (#41)
LDlite has been playing fast and loose with concatenating sql strings. It's probably fine but psycopg's sql module is intended to make this safe. While refactoring the existing code to implementing the sql module I've abstracted it into a database class intended to be overridden. In the first implementation we're still using the DBType enum and a single implementation, this will change in future into separate postgres and duckdb implementations without the DBType enum.