Refactor DBTypeDatabase into duckdb and postgres (#47)
In order to make fewer changes for the COPY FROM feature I created a single implementation of Database that utilized DBType. Now that sqlite support is gone I've created two implementations for duckdb and postgres. I wanted to use the dbapi types but duckdb doesn't quite play nicely with them. Doing this refactor makes the next few features easier to implement.

I was never happy with the iteration of records and ingestion of records. This moves the __id generation into the database logic away from the iteration logic. I also cleaned up the magic "first record is the total" behavior. As part of this I changed how tqdm was initialized to be closer to how I think it is intended to be used.