Add a system schema and load history table (#48)
In order to do the transformations more efficiently I'll need to create some custom functions. I've created a system schema for them to live in. Now that there's a place to put stuff I've also introduced a table recording load history for each table which will allow users and administrators to have a better understanding of how LDLite is running.

The test case setup is getting pretty gnarly and should maybe be reworked at some point. It would also be nice to have the End to End tests be run against postgres and duckdb.