Remove rollback from drop_tables logic (#49)
LDLite has always used exceptions for normal program logic around the dropping tables out of the catalog in case the catalog didn't exist. It leads to broken connections requiring rollbacks and I haven't been happy with the implementation of it as rollback works differently with pg and duckdb. This now checks for the catalog tables before trying to query them which cleans up the abstract base class.

As part of cleaning up the prefix object I removed the default schema from the key if there is no schema.